+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/dto/code/GatheringCode.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.gathering.shared.error.code;
+package com.sparta.moim.gathering.gathering.application.dto.code;
 
 import com.sparta.moim.common.response.Code;
 import lombok.Getter;
