+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/dto/event/GatheringAdminRevokeEvent.java
@@ -1,0 +1,8 @@
+package com.sparta.moim.gathering.gathering.application.dto.event;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record GatheringAdminRevokeEvent(UUID gatheringId, String ownerName) {
+}
