+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/dto/event/GatheringAdminSaveEvent.java
@@ -1,0 +1,8 @@
+package com.sparta.moim.gathering.gathering.application.dto.event;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record GatheringAdminSaveEvent(UUID gatheringId, String memberName, String type) {
+}
