--- gathering-service/src/main/java/com/sparta/moim/gathering/shared/dto/SharedGatheringRevokeMember.java
@@ -1,8 +1,0 @@
-package com.sparta.moim.gathering.shared.dto;
-
-import java.util.UUID;
-import lombok.Builder;
-
-@Builder
-public record SharedGatheringRevokeMember(UUID gatheringId, String ownerName) {
-}
